+++ classes/iterators/filters/recursives/extension.php
@@ -8,7 +8,7 @@ class extension extends \recursiveFilterIterator
 
 	public function __construct($mixed, array $acceptedExtensions = array(), \closure $iteratorFactory = null)
 	{
-		if ($mixed instanceof \recursiveIterator)
+		if ($mixed instanceof \RecursiveIterator)
 		{
 			parent::__construct($mixed);
 		}