+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # `dev-master`
 
+* [#624](https://github.com/atoum/atoum/pull/625) `atoum\iterators\recursives\directory\factory` is no longer a aggregated iterator ([@hywan])
+
 # 2.8.2 - 2016-08-12
 
 * [#620](https://github.com/atoum/atoum/pull/620) Add HTML coverage report from [reports extension](https://github.com/atoum/reports-extension) to AtoumTask for Phing ([@oxman])