+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # `dev-master`
 
-* [#624](https://github.com/atoum/atoum/pull/625) `atoum\iterators\recursives\directory\factory` is no longer a aggregated iterator ([@hywan])
+* [#625](https://github.com/atoum/atoum/pull/625) Isolation test execution engine is working on HHVM ([@hywan])
+* [#625](https://github.com/atoum/atoum/pull/625) `atoum\iterators\recursives\directory\factory` is no longer a aggregated iterator ([@hywan])
 
 # 2.8.2 - 2016-08-12