+++ CHANGELOG.md
@@ -17,6 +17,8 @@
 * [#605](https://github.com/atoum/atoum/pull/605) Handle `.autoloader.atoum.php` files to define tests autoloader ([@jubianchi])
 * [#605](https://github.com/atoum/atoum/pull/605) Add the `--autoloader-file`/`-af` CLI argument to define which autoloader file to user ([@jubianchi])
 * [#596](https://github.com/atoum/atoum/pull/596) Test methods' tags are inherited from test classes ([@jubianchi])
+* [#604](https://github.com/atoum/atoum/pull/604) Add a `addConfigurationCallable` method on the runner to allow extenion to register themselves ([@jubianchi], [@agallou])
+* [#604](https://github.com/atoum/atoum/pull/604) Add a `getExtension` method on the runner to configure an already loaded extension ([@jubianchi], [@agallou])
 
 # 2.7.0 - 2016-06-20